Aggregate AOD by Countries
This code aggregates the AOD data by countries, just as the other one
does for population. There are a few differences in this code, due to
the differences in the (folder) placement of uberrasters and source
raster, but the essence is the same. The code will generate many excel
tables (for all the years), and save them in the AOD data folder
(S:\particulates\data_processing\data\MODIS_AOD\generated\Temp_aggregate_aod_check).