dta2raster - preliminary
This code will take up any dta file that is in a given folder, and
transform it into a raster (it does everything down to creating the
raster). The input has to be ordered according to v1 (uber_code).
In the next version, I will construct a program and improve the
generality of this code.