Bringing the Wind data in
Some files that download the wind data.
IMPORTANT: Modification to rasters.py, please see CAREFULLY. Added a
possible line allowing for catalog aggregation of real-valued rasters
(the aggregation will take the mean of all the rasters in the catalog).
MAYBE, the type of raster should be an input (real x categorized)